refactor: validate 함수 통합
